Fix cargo packaging issue
Our build script was failing when running via `cargo package`. The build
script was removed to remedy this. This effects the error message
displayed when a consumer doesn't specify either hmac-openssl or
hmac-rust-crypto; now, the message is ambiguous. Documentation was
updated to help with that. The feature stuff should get a refactor in
the future so the build error doesn't occur in this library.